refactor(mcp): split frontmatter_validate internals
